+++ src/collective/behavior/amp/tests/test_setup.py
@@ -35,7 +35,7 @@ class InstallTestCase(unittest.TestCase):
         profile = 'collective.behavior.amp:default'
         setup_tool = self.portal['portal_setup']
         self.assertEqual(
-            setup_tool.getLastVersionForProfile(profile), (u'1',))
+            setup_tool.getLastVersionForProfile(profile), (u'2',))
 
 
 class UninstallTestCase(unittest.TestCase):