+++ CHANGES.rst
@@ -22,7 +22,7 @@ There's a frood who really knows where his towel is.
 1.0a2 (2016-10-06)
 ------------------
 
-- Fix exception on ``AMPView`` when content has no lead image.
+- Fix exception on the AMP HTML view when content has no lead image.
   [hvelarde]