+++ CHANGES.rst
@@ -6,6 +6,15 @@ There's a frood who really knows where his towel is.
 1.0a3 (unreleased)
 ------------------
 
+- Add class for social icon <div> tag.
+  [agnogueira]
+
+- Move description to article header.
+  [agnogueira]
+
+- Add missing script for <amp-analytics> component.
+  [agnogueira]
+
 - Do not fail if an image doesn't have a ``src`` attribute,
   or if it was referenced inside the body text as an external resource.
   [hvelarde]