+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 1.0a3 (unreleased)
 ------------------
 
+- Avoid exception when related items relation is broken.
+  [rodfersou]
+
 - Fix loading order for main AMP script.
   [hvelarde]