+++ CHANGES.rst
@@ -37,7 +37,8 @@ There's a frood who really knows where his towel is.
 - Fix various exceptions raised when content has Archetypes based lead images.
   [hvelarde]
 
-
+- Fix CSS error in logo position.
+  [kakshay21]
 1.0a2 (2016-10-06)
 ------------------