+++ CHANGES.rst
@@ -3,7 +3,7 @@ Changelog
 
 There's a frood who really knows where his towel is.
 
-1.0a3 (unreleased)
+1.0a3 (2017-04-13)
 ------------------
 
 - Fix i18n and update Brazilian Portuguese and Spanish translations (fixes `#44`_).